+++ dev-test/githunt/types.onlyEnums.ts
@@ -1,0 +1,16 @@
+/** A list of options for the sort order of the feed */
+export enum FeedType {
+  /** Sort by a combination of freshness and score, using Reddit's algorithm */
+  Hot = 'HOT',
+  /** Newest entries first */
+  New = 'NEW',
+  /** Highest score entries first */
+  Top = 'TOP',
+}
+
+/** The type of vote to record, when submitting a vote */
+export enum VoteType {
+  Cancel = 'CANCEL',
+  Down = 'DOWN',
+  Up = 'UP',
+}