+++ packages/plugins/other/visitor-plugin-common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @graphql-codegen/visitor-plugin-common
 
+## 2.8.0
+
+### Minor Changes
+
+- f1fb77bd4: feat: Add option to squash exactly similar fragment types
+
+## 2.7.6
+
+### Patch Changes
+
+- 9a5f31cb6: New option `onlyEnums` for Typescript
+
 ## 2.7.5
 
 ### Patch Changes