+++ packages/plugins/typescript/apollo-angular/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @graphql-codegen/typescript-apollo-angular
 
+## 3.4.10
+
+### Patch Changes
+
+- Updated dependencies [f1fb77bd4]
+  - @graphql-codegen/visitor-plugin-common@2.8.0
+
+## 3.4.9
+
+### Patch Changes
+
+- Updated dependencies [9a5f31cb6]
+  - @graphql-codegen/visitor-plugin-common@2.7.6
+
 ## 3.4.8
 
 ### Patch Changes