+++ packages/plugins/typescript/generic-sdk/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @graphql-codegen/typescript-generic-sdk
 
+## 2.3.10
+
+### Patch Changes
+
+- Updated dependencies [f1fb77bd4]
+  - @graphql-codegen/visitor-plugin-common@2.8.0
+
+## 2.3.9
+
+### Patch Changes
+
+- Updated dependencies [9a5f31cb6]
+  - @graphql-codegen/visitor-plugin-common@2.7.6
+
 ## 2.3.8
 
 ### Patch Changes