+++ packages/plugins/typescript/graphql-request/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @graphql-codegen/typescript-graphql-request
 
+## 4.4.8
+
+### Patch Changes
+
+- Updated dependencies [f1fb77bd4]
+  - @graphql-codegen/visitor-plugin-common@2.8.0
+
+## 4.4.7
+
+### Patch Changes
+
+- Updated dependencies [9a5f31cb6]
+  - @graphql-codegen/visitor-plugin-common@2.7.6
+
 ## 4.4.6
 
 ### Patch Changes