+++ packages/plugins/typescript/jit-sdk/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @graphql-codegen/typescript-jit-sdk
 
+## 2.0.9
+
+### Patch Changes
+
+- Updated dependencies [f1fb77bd4]
+  - @graphql-codegen/visitor-plugin-common@2.8.0
+
+## 2.0.8
+
+### Patch Changes
+
+- Updated dependencies [9a5f31cb6]
+  - @graphql-codegen/visitor-plugin-common@2.7.6
+
 ## 2.0.7
 
 ### Patch Changes