+++ packages/plugins/typescript/msw/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @graphql-codegen/typescript-msw
 
+## 1.0.10
+
+### Patch Changes
+
+- Updated dependencies [f1fb77bd4]
+  - @graphql-codegen/visitor-plugin-common@2.8.0
+
+## 1.0.9
+
+### Patch Changes
+
+- Updated dependencies [9a5f31cb6]
+  - @graphql-codegen/visitor-plugin-common@2.7.6
+
 ## 1.0.8
 
 ### Patch Changes