+++ packages/plugins/typescript/typescript/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @graphql-codegen/typescript
 
+## 2.4.11
+
+### Patch Changes
+
+- Updated dependencies [f1fb77bd4]
+  - @graphql-codegen/visitor-plugin-common@2.8.0
+
+## 2.4.10
+
+### Patch Changes
+
+- 9a5f31cb6: New option `onlyEnums` for Typescript
+- Updated dependencies [9a5f31cb6]
+  - @graphql-codegen/visitor-plugin-common@2.7.6
+
 ## 2.4.9
 
 ### Patch Changes