+++ packages/plugins/typescript/vue-apollo/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @graphql-codegen/typescript-vue-apollo
 
+## 3.2.12
+
+### Patch Changes
+
+- Updated dependencies [f1fb77bd4]
+  - @graphql-codegen/visitor-plugin-common@2.8.0
+
+## 3.2.11
+
+### Patch Changes
+
+- Updated dependencies [9a5f31cb6]
+  - @graphql-codegen/visitor-plugin-common@2.7.6
+
 ## 3.2.10
 
 ### Patch Changes